Make sure fs.change_dir is always checked and results in proper errors.

+++ src/luarocks/fetch/hg.lua
@@ -39,11 +39,13 @@ function get_sources(rockspec, extract, dest_dir)
    else
       store_dir = dest_dir
    end
-   fs.change_dir(store_dir)
+   local ok, err = fs.change_dir(store_dir)
+   if not ok then return nil, err end
    if not fs.execute(unpack(command)) then
       return nil, "Failed cloning hg repository."
    end
-   fs.change_dir(module)
+   local ok, err = fs.change_dir(module)
+   if not ok then return nil, err end
 
    fs.delete(dir.path(store_dir, module, ".hg"))
    fs.delete(dir.path(store_dir, module, ".hgignore"))
